refactor to mixin
