+++ addon/components/slds-control-group/checkbox.js
@@ -1,13 +1,9 @@
 import Ember from 'ember';
 import layout from '../../templates/components/slds-control-group/checkbox';
+import identity from '../../mixins/control-identity';
 
-export default Ember.Component.extend({
+export default Ember.Component.extend(identity, {
   layout,
   classNames: 'slds-form-element__control',
-  classNameBindings: ['isError:slds-has-error'],
-
-  formElementId: Ember.computed('elementId', function() {
-    let elementId = this.get('elementId');
-    return `${elementId}-field`;
-  })
+  classNameBindings: ['isError:slds-has-error']
 });
