+++ addon/components/slds-control.js
@@ -1,19 +1,10 @@
 import Ember from 'ember';
 import layout from '../templates/components/slds-control';
+import identity from '../mixins/control-identity';
 
-export default Ember.Component.extend({
+export default Ember.Component.extend(identity, {
   layout,
 
   classNames: 'slds-form-element',
-  classNameBindings: ['isError:slds-has-error'],
-
-  formElementId: Ember.computed('elementId', function() {
-    let elementId = this.get('elementId');
-    return `${elementId}-field`;
-  }),
-
-  errorElementId: Ember.computed('elementId', function() {
-    let elementId = this.get('elementId');
-    return `${elementId}-error`;
-  })
+  classNameBindings: ['isError:slds-has-error']
 });
